+++ lib/src/packet/contained_packet.dart
@@ -2,6 +2,7 @@
 // For the full copyright and license information, please view the LICENSE
 // file that was distributed with this source code.
 
+import 'dart:math';
 import 'dart:typed_data';
 
 import '../enum/packet_tag.dart';
@@ -9,6 +10,9 @@ import '../crypto/math/int_ext.dart';
 
 /// Author Nguyen Van Nguyen <nguyennv1981@gmail.com>
 abstract class ContainedPacket {
+  static const partialMinSize = 512;
+  static const partialMaxSize = 1024;
+
   final PacketTag tag;
 
   ContainedPacket(this.tag);
@@ -18,22 +22,55 @@ abstract class ContainedPacket {
 
   /// Serializes packet to bytes
   Uint8List encode() {
-    final bodyData = toByteData();
-    final bodyLen = bodyData.length;
-
-    final List<int> headerData;
-    final hdr = 0x80 | 0x40 | tag.value;
-    if (bodyLen < 192) {
-      headerData = [hdr, bodyLen];
-    } else if (bodyLen <= 8383) {
-      headerData = [
-        hdr,
-        (((bodyLen - 192) >> 8) & 0xff) + 192,
-        bodyLen - 192,
-      ];
-    } else {
-      headerData = [hdr, 0xff, ...bodyLen.pack32()];
+    switch (tag) {
+      case PacketTag.aeadEncryptedData:
+      case PacketTag.compressedData:
+      case PacketTag.literalData:
+      case PacketTag.symEncryptedData:
+      case PacketTag.symEncryptedIntegrityProtectedData:
+        return _partialEncode();
+      default:
+        final bodyData = toByteData();
+        return Uint8List.fromList([
+          0xc0 | tag.value,
+          ..._simpleLength(bodyData.length),
+          ...bodyData,
+        ]);
     }
-    return Uint8List.fromList([...headerData, ...bodyData]);
+  }
+
+  /// Encode package to the openpgp partial body specifier
+  Uint8List _partialEncode() {
+    final List<int> partialData = [];
+    var bodyData = toByteData();
+    var dataLengh = bodyData.length;
+    while (dataLengh >= partialMinSize) {
+      final maxSize = min(partialMaxSize, dataLengh);
+      final powerOf2 = min((log(maxSize) / ln2).toInt(), 30);
+      final chunkSize = 1 << powerOf2;
+      partialData.addAll(
+        [
+          224 + powerOf2,
+          ...bodyData.sublist(0, chunkSize),
+        ],
+      );
+      bodyData = bodyData.sublist(chunkSize);
+      dataLengh = bodyData.length;
+    }
+    partialData.addAll([
+      ..._simpleLength(dataLengh),
+      ...bodyData,
+    ]);
+    return Uint8List.fromList([0xc0 | tag.value, ...partialData]);
+  }
+
+  List<int> _simpleLength(int length) {
+    if (length < 192) {
+      return [length];
+    } else if (length < 8384) {
+      return [(((length - 192) >> 8) & 0xff) + 192, length - 192];
+    } else {
+      return [0xff, ...length.pack32()];
+    }
   }
 }