+++ lib/src/packet/secret_key.dart
@@ -7,6 +7,7 @@ import 'dart:typed_data';
 import 'package:pointycastle/export.dart';
 
 import '../crypto/math/byte_ext.dart';
+import '../crypto/math/int_ext.dart';
 import '../crypto/symmetric/base_cipher.dart';
 import '../enum/curve_info.dart';
 import '../enum/dh_key_size.dart';
@@ -80,24 +81,29 @@ class SecretKeyPacket extends ContainedPacket implements KeyPacket {
     }
 
     Uint8List? iv;
-    if (!(s2k != null && s2k.type == S2kType.gnu) &&
-        s2kUsage != S2kUsage.none) {
+    if (!(s2k != null && s2k.type == S2kType.gnu) && s2kUsage != S2kUsage.none) {
       final blockSize = symmetric.blockSize;
       iv = bytes.sublist(pos, pos + blockSize);
       pos += blockSize;
     }
 
     KeyParams? secretParams;
+    var keyData = bytes.sublist(pos);
     if (s2kUsage == S2kUsage.none) {
+      final checksum = keyData.sublist(keyData.length - 2);
+      keyData = keyData.sublist(0, keyData.length - 2);
+      if (!checksum.equals(_computeChecksum(keyData))) {
+        throw StateError('Key checksum mismatch!');
+      }
       secretParams = _parseSecretParams(
-        bytes.sublist(pos),
+        keyData,
         publicKey.algorithm,
       );
     }
 
     return SecretKeyPacket(
       publicKey,
-      bytes.sublist(pos),
+      keyData,
       s2kUsage: s2kUsage,
       symmetric: symmetric,
       s2k: s2k,
@@ -239,9 +245,7 @@ class SecretKeyPacket extends ContainedPacket implements KeyPacket {
     if (secretParams == null) {
       final Uint8List clearText;
       if (isEncrypted) {
-        final key =
-            await s2k?.produceKey(passphrase, symmetric.keySizeInByte) ??
-                Uint8List(symmetric.keySizeInByte);
+        final key = await s2k?.produceKey(passphrase, symmetric.keySizeInByte) ?? Uint8List(symmetric.keySizeInByte);
         final cipher = BufferedCipher(symmetric.cfbCipherEngine)
           ..init(
             false,
@@ -295,8 +299,7 @@ class SecretKeyPacket extends ContainedPacket implements KeyPacket {
       return keyParams.validatePublicParams(publicParams as DSAPublicParams);
     }
     if (keyParams is ElGamalSecretParams) {
-      return keyParams
-          .validatePublicParams(publicParams as ElGamalPublicParams);
+      return keyParams.validatePublicParams(publicParams as ElGamalPublicParams);
     }
     if (keyParams is ECSecretParams) {
       return keyParams.validatePublicParams(publicParams as ECPublicParams);
@@ -328,6 +331,7 @@ class SecretKeyPacket extends ContainedPacket implements KeyPacket {
         ...publicKey.toByteData(),
         S2kUsage.none.value,
         ...keyData,
+        ..._computeChecksum(keyData),
       ]);
     }
   }
@@ -355,5 +359,13 @@ class SecretKeyPacket extends ContainedPacket implements KeyPacket {
           'Public key algorithm ${algorithm.name} is unsupported.',
         );
     }
+  }
+
+  static Uint8List _computeChecksum(Uint8List keyData) {
+    var sum = 0;
+    for (var i = 0; i < keyData.length; i++) {
+      sum = (sum + keyData[i]) & 0xffff;
+    }
+    return sum.pack16();
   }
 }